refactor: consolidate renderer registries and optimize BEWLR
- Move all renderer registrations to single RendererRegistry class
- Convert FlyingBroomItemRenderer to singleton pattern to reduce memory usage
- Remove old entity model registrations that were causing unused OBJ loading
- Clean up scattered client init files
