+++ src/main/java/org/sosly/witchcraft/items/FlyingBroomItem.java
@@ -53,16 +53,9 @@ public class FlyingBroomItem extends TieredItem {
     @Override
     public void initializeClient(Consumer<IClientItemExtensions> consumer) {
         consumer.accept(new IClientItemExtensions() {
-            private final NonNullLazy<BlockEntityWithoutLevelRenderer> renderer = NonNullLazy.of(() -> 
-                new FlyingBroomItemRenderer(
-                    Minecraft.getInstance().getBlockEntityRenderDispatcher(),
-                    Minecraft.getInstance().getEntityModels()
-                )
-            );
-
             @Override
             public BlockEntityWithoutLevelRenderer getCustomRenderer() {
-                return renderer.get();
+                return FlyingBroomItemRenderer.getInstance();
             }
         });
     }
